refactor : test 파일 이동 및 모바일 반응형 작업 초기 세팅
